+++ src/main/java/com/aluminium/acoe/com/controller/UserController.java
@@ -1,7 +1,7 @@
-package com.aluminium.acoe.controller;
+package com.aluminium.acoe.com.controller;
 
-import com.aluminium.acoe.dto.UserDto;
-import com.aluminium.acoe.service.UserService;
+import com.aluminium.acoe.com.dto.UserDto;
+import com.aluminium.acoe.com.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
